Break out loop to different module
